Don't create beta information for un-expanded functions.
Currently the meta information for a dataset is created whenever it is
created.  This means when functions are bound and expanded the meta
information can be created lots of times.

Unfortunately this fix doesn't quite work.  See dab50 for an example
where an alien data type causes problems and private:insure2.eclxml
for a failure to find a function correctly.

Branch released to revisit later and see if the issues can be resolved.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>